// packages/sirius-web/frontend/sirius-web-application/src/omnibox/PublishStudioOmniboxCommandOverride.types.ts
/*******************************************************************************
 * Copyright (c) 2025 Obeo.
 * This program and the accompanying materials
 * are made available under the terms of the Eclipse Public License v2.0
 * which accompanies this distribution, and is available at
 * https://www.eclipse.org/legal/epl-2.0/
 *
 * SPDX-License-Identifier: EPL-2.0
 *
 * Contributors:
 *     Obeo - initial API and implementation
 *******************************************************************************/

export interface PublishStudioState {
  version: string;
  versionIsInvalid: boolean;
  pristine: boolean;
  description: string;
}